fix: predicates read nodes from scheduler cache
