+++ pkg/scheduler/internal/cache/fake/fake_cache.go
@@ -104,3 +104,8 @@ func (c *Cache) Snapshot() *internalcache.Snapshot {
 
 // NodeTree is a fake method for testing.
 func (c *Cache) NodeTree() *internalcache.NodeTree { return nil }
+
+// GetNodeInfo is a fake method for testing.
+func (c *Cache) GetNodeInfo(nodeName string) (*v1.Node, error) {
+	return nil, nil
+}
